+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.3.0
 
- Enhancements
+Enhancements
 
  - Added `AudioDeviceSelector.VERSION` constant so developers can access the version of AudioSwitch at runtime.
  - Added `AudioDeviceSelector.loggingEnabled` property so developers can configure AudioSwitch logging behavior at runtime. By default, AudioSwitch logging is disabled. Reference the following snippet to enable AudioSwitch logging: